+++ pineappl_cli/tests/convolve.rs
@@ -2,7 +2,7 @@ use assert_cmd::Command;
 
 const HELP_STR: &str = "Convolutes a PineAPPL grid with a PDF set
 
-Usage: pineappl convolute [OPTIONS] <INPUT> <PDFSETS>...
+Usage: pineappl convolve [OPTIONS] <INPUT> <PDFSETS>...
 
 Arguments:
   <INPUT>       Path of the input grid
@@ -168,7 +168,7 @@ For more information, try '--help'.
 fn help() {
     Command::cargo_bin("pineappl")
         .unwrap()
-        .args(["convolute", "--help"])
+        .args(["convolve", "--help"])
         .assert()
         .success()
         .stdout(HELP_STR);
@@ -179,8 +179,7 @@ fn default() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed",
         ])
@@ -195,8 +194,7 @@ fn force_positive() {
         .unwrap()
         .args([
             "--force-positive",
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed",
         ])
@@ -210,8 +208,7 @@ fn default_multiple_pdfs() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed",
             "324900=NNPDF31_nlo_as_0118_luxqed",
@@ -226,8 +223,7 @@ fn multiple_pdfs_with_new_construction() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed/0",
             "NNPDF31_nlo_as_0118_luxqed/1",
@@ -242,8 +238,7 @@ fn multiple_pdfs_with_relabeling() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed",
             "NNPDF31_nlo_as_0118_luxqed/1=other mc=1.4",
@@ -258,8 +253,7 @@ fn two_pdfs_with_order_subset() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "--orders=a2",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed/0",
@@ -275,8 +269,7 @@ fn three_pdfs() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed/0",
             "NNPDF31_nlo_as_0118_luxqed/1",
@@ -292,8 +285,7 @@ fn bins_13567() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "--bins=1,3,5-7",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed",
@@ -308,8 +300,7 @@ fn integrated() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "--integrated",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed",
@@ -324,8 +315,7 @@ fn integrated_multiple_pdfs() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "--integrated",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed",
@@ -341,8 +331,7 @@ fn orders_a2_a3() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "--orders=a2,a3",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed",
@@ -357,8 +346,7 @@ fn wrong_orders() {
     Command::cargo_bin("pineappl")
         .unwrap()
         .args([
-            "--silence-lhapdf",
-            "convolute",
+            "convolve",
             "--orders=a2a2as2",
             "../test-data/LHCB_WP_7TEV.pineappl.lz4",
             "NNPDF31_nlo_as_0118_luxqed",