+++ CHANGELOG.md
@@ -15,6 +15,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- changed the official name of the CLI subcommand `convolute` to `convolve`,
+  because the latter is the proper verb of 'convolution'. The old name
+  `convolute` is now an alias of `convolve`, which means both can be used. The
+  methods `Grid::convolute*` are left unchanged and will be renamed in later
+  version
+- changed switch `--silence-lhapdf` to `--lhapdf-banner` and suppress LHAPDF's
+  banners by default, unless `--lhapdf-banner` is given
 - `Grid::evolve` has now been marked deprecated
 - switched from `lhapdf` to `managed-lhapdf` crate which automatically
   downloads PDF sets when they are needed